% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccer.r
\name{bayesian_linear_indep}
\alias{bayesian_linear_indep}
\title{Bayesian estimator for linear with binary classification as \emph{independent} variable}
\usage{
bayesian_linear_indep(
  formula,
  data,
  fn = fn,
  fp = fp,
  tp = tp,
  tn = tn,
  npv = npv,
  ppv = ppv,
  predicted_col_name = "xeb",
  tau_prior = 0.001,
  working_dir = "",
  n.iter = n.iter
)
}
\arguments{
\item{formula}{R formula, such as y ~ x}

\item{data}{a data.frame object}

\item{fn}{number of false negatives in validation data}

\item{fp}{number of false positives in validation data}

\item{tp}{number of true positives in validation data}

\item{tn}{number of true negatives in validation data}

\item{predicted_col_name}{string or integer indicating the column of the variable that is predicted}

\item{tau_prior}{prior values of hyperparameters of Bayesian regression variance}

\item{working_dir}{directory to store JAGS temporary files}

\item{n.iter}{number of Bayesian draws}

\item{fnr}{false negative rate}

\item{fpr}{false positive rate}
}
\value{
a dataframe contining estimated values, standard errors, z values and p values
}
\description{
This functions takes binary classification as \emph{independent} variables, apply Bayesian estimator, and returned corrected coefficient estimates.
}
