% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccer.r
\name{bayesian_logistic}
\alias{bayesian_logistic}
\title{Bayesian estimator for logistic regression with binary classification as \emph{outcome} variable}
\usage{
bayesian_logistic(
  formula,
  data,
  fn = fn,
  fp = fp,
  tp = tp,
  tn = tn,
  fpr = fpr,
  fnr = fnr,
  tau_prior = 0.001,
  working_dir = "",
  n.iter = n.iter
)
}
\arguments{
\item{formula}{R formula, such as y ~ x}

\item{data}{a data.frame object}

\item{fn}{number of false negatives in validation data}

\item{fp}{number of false positives in validation data}

\item{tp}{number of true positives in validation data}

\item{tn}{number of true negatives in validation data}

\item{fpr}{false positive rate}

\item{fnr}{false negative rate}

\item{tau_prior}{prior values of hyperparameters of Bayesian regression variance}

\item{working_dir}{directory to store JAGS temporary files}

\item{n.iter}{number of Bayesian draws}
}
\value{
a dataframe contining estimated values, standard errors, z values and p values
}
\description{
This functions takes binary classification as \emph{outcome} variables, apply Bayesianestimator, and returned corrected coefficient estimates.
}
