% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccer.r
\name{MLE_linear_indep_bootstrap}
\alias{MLE_linear_indep_bootstrap}
\title{Bootstrap estimator for linear regression with binary classification as \emph{independent} variables}
\usage{
MLE_linear_indep_bootstrap(
  formula,
  data,
  predicted_col_name,
  tl,
  pl,
  nboot = 10000,
  return_boot_samples = FALSE
)
}
\arguments{
\item{formula}{R formula, such as y ~ x}

\item{data}{a data.frame object}

\item{predicted_col_name}{string or integer indicating the column of the variable that is predicted}

\item{tl}{vector of true values}

\item{pl}{vector of predicted values; needs to be of the same dimension of the vevtor of true values.}

\item{nboot}{number of bootstrap samples}
}
\value{
_boot_samples whether to return raw bootstrap estimates

if return_boot_samples is false, return mean/standard error/quantile statistics from bootstrap sample only; otherwise return a list, the first element of which is mean/standard error/quantile statistics from bootstrap sample , and the second ele,ent of which is \code{nboot} samples from bootstrap draws
}
\description{
This functions takes binary classification as \emph{independent} variables, apply MLE estimator, and returned corrected coefficient estimates.
}
