% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccer.r
\name{MLE_LPM}
\alias{MLE_LPM}
\title{MLE estimator for linear regression with binary classification as \emph{outcome} variable (linear probability model)}
\usage{
MLE_LPM(formula, data, fnr = 0, fpr = 0)
}
\arguments{
\item{formula}{R formula, such as y ~ x}

\item{data}{a data.frame object}

\item{fnr}{false negative rate}

\item{fpr}{false positive rate}
}
\value{
a dataframe contining estimated values, standard errors, z values and p values
}
\description{
This functions takes binary classification as \emph{outcome} variables, apply MLE estimator, and returned corrected coefficient estimates.
}
